+++ demo/angular/src/app/container/container.component.ts
@@ -17,7 +17,6 @@ export class ContainerComponent implements OnInit {
         },
         fpsLimit: 60,
         interactivity: {
-            detectsOn: 'window',
             events: {
                 onClick: {
                     enable: true,