+++ demo/ionic/src/app/tab1/tab1.page.ts
@@ -21,7 +21,6 @@ export class Tab1Page {
     },
     fpsLimit: 60,
     interactivity: {
-      detectsOn: 'window',
       events: {
         onClick: {
           enable: true,