+++ interactions/particles/collisions/src/index.ts
@@ -1,6 +1,6 @@
 import type { Main } from "tsparticles-engine";
 import { Collider } from "./Collider";
 
-export function loadParticlesCollisionsInteraction(tsParticles: Main): void {
-    tsParticles.addInteractor("particlesCollisions", (container) => new Collider(container));
+export async function loadParticlesCollisionsInteraction(tsParticles: Main): Promise<void> {
+    await tsParticles.addInteractor("particlesCollisions", (container) => new Collider(container));
 }