+++ interactions/particles/links/src/plugin.ts
@@ -21,8 +21,8 @@ class Index implements IPlugin {
     }
 }
 
-export function loadPlugin(tsParticles: Main): void {
+export async function loadPlugin(tsParticles: Main): Promise<void> {
     const plugin = new Index();
 
-    tsParticles.addPlugin(plugin);
+    await tsParticles.addPlugin(plugin);
 }