+++ shapes/circle/src/index.ts
@@ -1,6 +1,6 @@
 import type { Main } from "tsparticles-engine";
 import { CircleDrawer } from "./CircleDrawer";
 
-export function loadCircleShape(tsParticles: Main): void {
-    tsParticles.addShape("circle", new CircleDrawer());
+export async function loadCircleShape(tsParticles: Main): Promise<void> {
+    await tsParticles.addShape("circle", new CircleDrawer());
 }