+++ updaters/life/src/index.ts
@@ -1,6 +1,6 @@
 import type { Main } from "tsparticles-engine";
 import { LifeUpdater } from "./LifeUpdater";
 
-export function loadLifeUpdater(tsParticles: Main): void {
-    tsParticles.addParticleUpdater("life", (container) => new LifeUpdater(container));
+export async function loadLifeUpdater(tsParticles: Main): Promise<void> {
+    await tsParticles.addParticleUpdater("life", (container) => new LifeUpdater(container));
 }