+++ updaters/opacity/src/index.ts
@@ -1,6 +1,6 @@
 import type { Main } from "tsparticles-engine";
 import { OpacityUpdater } from "./OpacityUpdater";
 
-export function loadOpacityUpdater(tsParticles: Main): void {
-    tsParticles.addParticleUpdater("opacity", (container) => new OpacityUpdater(container));
+export async function loadOpacityUpdater(tsParticles: Main): Promise<void> {
+    await tsParticles.addParticleUpdater("opacity", (container) => new OpacityUpdater(container));
 }