+++ updaters/outModes/src/index.ts
@@ -1,6 +1,6 @@
 import type { Main } from "tsparticles-engine";
 import { OutOfCanvasUpdater } from "./OutOfCanvasUpdater";
 
-export function loadOutModesUpdater(tsParticles: Main): void {
-    tsParticles.addParticleUpdater("outModes", (container) => new OutOfCanvasUpdater(container));
+export async function loadOutModesUpdater(tsParticles: Main): Promise<void> {
+    await tsParticles.addParticleUpdater("outModes", (container) => new OutOfCanvasUpdater(container));
 }