+++ updaters/tilt/src/index.ts
@@ -1,6 +1,6 @@
 import type { Main } from "tsparticles-engine";
 import { TiltUpdater } from "./TiltUpdater";
 
-export function loadTiltUpdater(tsParticles: Main): void {
-    tsParticles.addParticleUpdater("tilt", (container) => new TiltUpdater(container));
+export async function loadTiltUpdater(tsParticles: Main): Promise<void> {
+    await tsParticles.addParticleUpdater("tilt", (container) => new TiltUpdater(container));
 }