+++ demo/vue3/CHANGELOG.md
@@ -4,11 +4,14 @@ All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
 # [2.0.0-beta.2](https://github.com/matteobruni/tsparticles/compare/particles.vue3-demo@1.18.4...particles.vue3-demo@2.0.0-beta.2) (2021-10-06)
+# [1.20.0](https://github.com/matteobruni/tsparticles/compare/particles.vue3-demo@1.19.0...particles.vue3-demo@1.20.0) (2021-10-28)
 
 
 ### Features
 
 * splitting engine from slim and full bundles (v2) ([268b78c](https://github.com/matteobruni/tsparticles/commit/268b78c12d6c54069893d27643cfe7a30f3be777))
+* optmize demo of vue3 ([88664ef](https://github.com/matteobruni/tsparticles/commit/88664ef4dea5a9bd42c12a8cc2975b4fea78d4a4))
+* optmize demo of vue3 ([6618f5d](https://github.com/matteobruni/tsparticles/commit/6618f5ddd384fa4b3ca3e984033e45691a1ed986))