+++ plugins/infection/CHANGELOG.md
@@ -15,6 +15,14 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 # [2.0.0-beta.1](https://github.com/matteobruni/tsparticles/compare/tsparticles-plugin-infection@2.0.0-beta.0...tsparticles-plugin-infection@2.0.0-beta.1) (2021-10-06)
+# [1.37.0](https://github.com/matteobruni/tsparticles/compare/tsparticles-plugin-infection@1.36.0...tsparticles-plugin-infection@1.37.0) (2021-10-28)
+
+**Note:** Version bump only for package tsparticles-plugin-infection
+
+
+
+
+
 # [1.36.0](https://github.com/matteobruni/tsparticles/compare/tsparticles-plugin-infection@1.35.4...tsparticles-plugin-infection@1.36.0) (2021-10-14)
 
 **Note:** Version bump only for package tsparticles-plugin-infection