+++ presets/seaAnemone/CHANGELOG.md
@@ -16,6 +16,14 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 # [2.0.0-beta.1](https://github.com/matteobruni/tsparticles/compare/tsparticles-preset-sea-anemone@2.0.0-beta.0...tsparticles-preset-sea-anemone@2.0.0-beta.1) (2021-10-06)
+# [1.22.0](https://github.com/matteobruni/tsparticles/compare/tsparticles-preset-sea-anemone@1.21.0...tsparticles-preset-sea-anemone@1.22.0) (2021-10-28)
+
+**Note:** Version bump only for package tsparticles-preset-sea-anemone
+
+
+
+
+
 # [1.21.0](https://github.com/matteobruni/tsparticles/compare/tsparticles-preset-sea-anemone@1.20.4...tsparticles-preset-sea-anemone@1.21.0) (2021-10-14)
 
 **Note:** Version bump only for package tsparticles-preset-sea-anemone